lifecycletest: Move nop-lifecycle into test package
