+++ yarpctest/fake_peer_list.go
@@ -26,7 +26,7 @@ import (
 
 	"go.uber.org/yarpc/api/peer"
 	"go.uber.org/yarpc/api/transport"
-	"go.uber.org/yarpc/pkg/lifecycle"
+	"go.uber.org/yarpc/pkg/lifecycletest"
 )
 
 // FakePeerListOption is an option for NewFakePeerList.
@@ -49,7 +49,7 @@ type FakePeerList struct {
 // NewFakePeerList returns a fake peer list.
 func NewFakePeerList(opts ...FakePeerListOption) *FakePeerList {
 	pl := &FakePeerList{
-		Lifecycle: lifecycle.NewNop(),
+		Lifecycle: lifecycletest.NewNop(),
 	}
 	for _, opt := range opts {
 		opt(pl)
