+++ yarpctest/fake_transport.go
@@ -24,7 +24,7 @@ import (
 	"go.uber.org/yarpc/api/peer"
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/peer/hostport"
-	"go.uber.org/yarpc/pkg/lifecycle"
+	"go.uber.org/yarpc/pkg/lifecycletest"
 )
 
 // FakeTransportOption is an option for NewFakeTransport.
@@ -41,7 +41,7 @@ func NopTransportOption(nopOption string) FakeTransportOption {
 // NewFakeTransport returns a fake transport.
 func NewFakeTransport(opts ...FakeTransportOption) *FakeTransport {
 	t := &FakeTransport{
-		Lifecycle: lifecycle.NewNop(),
+		Lifecycle: lifecycletest.NewNop(),
 	}
 	for _, opt := range opts {
 		opt(t)
